requirements.txt
Flask==1.1.2
google-oauth==1.0.1
google-cloud-dialogflow==2.1.2
Flask-Caching==1.9.0
flask-socketio==5.0.1
pyTelegramBotAPI==3.7.7